moar SideEffect + rename to InitialLoading
