+++ androidApp/src/main/java/dev/johnoreilly/confetti/initial_loading/navigation/SplashNavigation.kt
@@ -1,27 +1,27 @@
-package dev.johnoreilly.confetti.splash.navigation
+package dev.johnoreilly.confetti.initial_loading.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import dev.johnoreilly.confetti.conferences.navigation.ConferencesKey
 import dev.johnoreilly.confetti.sessions.navigation.SessionsKey
-import dev.johnoreilly.confetti.splash.SplashRoute
+import dev.johnoreilly.confetti.initial_loading.InitialLoadingRoute
 
-private const val base = "splash"
+private const val base = "initial_loading"
 
 private val pattern = base
 
-object SplashKey {
+object InitialLoadingKey {
     val route = base
 }
 
-fun NavGraphBuilder.splashGraph(
+fun NavGraphBuilder.initialLoadingGraph(
     navigateToConferences: (ConferencesKey) -> Unit,
     navigateToSessions: (SessionsKey) -> Unit
 ) {
     composable(
         route = pattern,
     ) {
-        SplashRoute(
+        InitialLoadingRoute(
             navigateToConferences = navigateToConferences,
             navigateToSessions = navigateToSessions
         )
